+++ SistemaDeNominas.BD.BluePrint/schema/Puestos.sql
@@ -3,6 +3,6 @@
 	[Id] INT  NOT NULL PRIMARY KEY, 
     [Nombre] VARCHAR(64) NOT NULL, 
     [NivelDeRiesgo] SMALLINT NOT NULL, 
-    [MinimoSalario] MONEY NOT NULL DEFAULT 14161.00, 
-    [MaximoSalario] INT NOT NULL
+    [MinimoSalario] MONEY NOT NULL DEFAULT 14161.00,
+    [MaximoSalario] MONEY NOT NULL
 )
